+++ ModdedMinecraftClub.MemberBot/ModdedMinecraftClub.MemberBot.Bot/Helper.cs
@@ -1,5 +1,9 @@
 ﻿using System;
 using System.Globalization;
+using System.IO;
+using ModdedMinecraftClub.MemberBot.Bot.Models;
+using YamlDotNet.Serialization;
+using YamlDotNet.Serialization.NamingConventions;
 
 namespace ModdedMinecraftClub.MemberBot.Bot
 {
@@ -10,6 +14,18 @@ namespace ModdedMinecraftClub.MemberBot.Bot
             var config = Program.Config.Mysql;
             
             return $"Server={config.ServerIp};Port={config.Port};Database={config.DatabaseName};Uid={config.Username};Pwd={config.Password};Allow User Variables=True";
+        }
+        
+        public static ConfigRoot LoadConfigFile()
+        {
+            var currentDir = Directory.GetCurrentDirectory();
+            var mainDir = Directory.GetParent(currentDir).ToString();
+            var path = Path.Combine(mainDir, "config.yml");
+            var deserializer = new DeserializerBuilder()
+                .WithNamingConvention(new CamelCaseNamingConvention())
+                .Build();
+            
+            return deserializer.Deserialize<ConfigRoot>(File.ReadAllText(path));
         }
         
         public static DateTime NormalizeDate(string originalJsonDate)
