+++ ModdedMinecraftClub.MemberBot/ModdedMinecraftClub.MemberBot.Bot/ModdedMinecraftClub.MemberBot.Bot.csproj
@@ -8,12 +8,13 @@
     </PropertyGroup>
 
     <ItemGroup>
-      <PackageReference Include="ConsoleTables" Version="2.3.0" />
-      <PackageReference Include="Dapper" Version="2.0.4" />
-      <PackageReference Include="Discord.Net" Version="2.1.1" />
-      <PackageReference Include="MySql.Data" Version="8.0.19" />
-      <PackageReference Include="Newtonsoft.Json" Version="12.0.3-beta1" />
-      <PackageReference Include="YamlDotNet" Version="6.1.2" />
+      <PackageReference Include="ConsoleTables" Version="2.4.1" />
+      <PackageReference Include="Dapper" Version="2.0.35" />
+      <PackageReference Include="Discord.Net" Version="2.2.0" />
+      <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="5.0.0-preview.4.20251.6" />
+      <PackageReference Include="MySql.Data" Version="8.0.20" />
+      <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
+      <PackageReference Include="YamlDotNet" Version="8.1.2" />
     </ItemGroup>
     
 </Project>
